+++ 404.php
@@ -1,54 +1,19 @@
 <?php
 /**
- * The template for displaying 404 pages (Not Found).
+ * 404 Template 
+ * Description: Displays content for pages not found.
  *
  * @package WordPress
  * @subpackage Most
  */
-get_header(); ?>
-<div class="container">
-    <div class="row">
-        <div class="span12"><?php
-            if (function_exists('most_breadcrumbs')) {
-                most_breadcrumbs();
-            } ?>
-        </div>
-    </div>
-    <div class="row content">
-        <div class="span8">
-            <header class="page-title">
-                <h1><?php _e('This is Embarrassing', 'moux'); ?></h1>
-            </header>
-            <p class="lead"><?php
-            _e('It seems we can&rsquo;t find what you&rsquo;re looking for. Perhaps searching, or one of the links below, can help.','moux'); ?>
-            </p>
-           <div class="well">
-               <?php get_search_form(); ?>
-           </div>
-           <div class="row">
-               <div class="span4">
-                   <h2>All Pages</h2>
-                   <?php wp_page_menu(); ?>
-               </div>
-               <!--/.span4 -->
-               <div class="span4">
-                   <?php the_widget('WP_Widget_Recent_Posts'); ?>
-                   <h2><?php _e('Most Used Categories', 'moux'); ?></h2>
-                   <ul><?php
-                        wp_list_categories(
-                            array(
-                               'orderby' => 'count',
-                               'order' => 'DESC',
-                               'show_count' => 1,
-                               'title_li' => '',
-                               'number' => 10
-                            )
-                        ); ?>
-                    </ul>
-                </div>
-                <!--/.span4 -->
-            </div>
-            <!--/.row -->
-        </div><?php
-    get_sidebar();
+get_header();
+   get_sidebar('left'); ?>
+   <div id="404-content" class="span6">
+      <h2><?php _e('Page Not Found', 'most'); ?></h2>
+      <p class="lead">It seems we can't find what you're looking for. Perhaps searching keywords or clicking one of the links below can help.</p>
+      <div class="well">
+         <?php get_search_form(); ?>
+      </div>
+   </div><?php
+   get_sidebar('right');
 get_footer();