+++ front-page.php
@@ -6,18 +6,7 @@
  * @subpackage Most
  */
 get_header();
-?>
-<div class="container">
-	<aside id="sidebar-left" class="span3"><?php
-        wp_nav_menu( array(
-            'theme_location' => 'side-menu',
-            'container' => 'false',
-            'menu_class' => 'nav nav-list',
-            'menu_id' => 'side-menu',
-            'fallback_cb' => ''
-        ) );
-        get_sidebar('left'); ?>
-	</aside>
+	get_sidebar('left'); ?>
 	<section id="front-content" class="span6">
 		<h2 id="today-date">Today at the MOST - <?php echo get_todays_date('l, F j, Y'); ?></h2>
 		<h3 id="today-hours">Hours of Operation Today: <?php echo get_most_hours(); ?></h3><?php
@@ -33,10 +22,6 @@ get_header();
 		elseif ( get_most_hours()!='closed' && ( $events->found_posts>0 || $shows->found_posts>0 ) ) :
 			get_template_part('template-parts/events');
 		endif; ?>
-	</section>
-	<aside id="sidebar-right" class="span3 pull-right"><?php
-        get_sidebar('right'); ?>
-	</aside>
-</div><!--/.container-->
-<?php
+	</section><?php
+	get_sidebar('right');
 get_footer(); ?>