+++ page.php
@@ -1,20 +1,19 @@
 <?php
 /**
  * Default Template
- * Description: Page template with a content container and right sidebar.
+ * Description: Displays a content container with a left and right sidebar.
  *
  * @package WordPress
  * @subpackage Most
  */
 get_header();
-while ( have_posts() ) : the_post(); ?>
-    <div class="container">
-        <h1><?php the_title();?></h1>
-        <div class="span8"><?php
+	get_sidebar('left');
+	while ( have_posts() ) : the_post(); ?>
+	    <div class="span6">
+	        <h2><?php the_title();?></h2><?php
             the_content();
             edit_post_link(__('Edit', 'most'), '<span class="edit-link">', '</span>'); ?>
-        </div><!-- /.span8 -->
-    </div><?php
-endwhile; // end of the loop.
-//get_sidebar();
+	    </div><?php
+	endwhile; // end of the loop.
+	get_sidebar('right');
 get_footer(); ?>