+++ sidebar-left.php
@@ -5,6 +5,16 @@
  * @package WordPress
  * @subpackage Most
  */
-if (function_exists('dynamic_sidebar')) {
-    dynamic_sidebar('left-sidebar');
-}
+?>
+<aside id="sidebar-left" class="span3"><?php
+    wp_nav_menu( array(
+        'theme_location' => 'side-menu',
+        'container' => 'false',
+        'menu_class' => 'nav nav-list',
+        'menu_id' => 'side-menu',
+        'fallback_cb' => ''
+    ) );
+    if (function_exists('dynamic_sidebar')) {
+	    dynamic_sidebar('left-sidebar');
+	} ?>
+</aside>