+++ single.php
@@ -7,14 +7,14 @@
  * @subpackage Most
  */
 get_header();
-while ( have_posts() ) : the_post(); ?>
-    <div class="container">
-        <h2><?php the_title(); ?></h2>
-        <div class="span8"><?php
+	get_sidebar('left');
+	while ( have_posts() ) : the_post(); ?>
+	    <div class="span6">
+	        <h2><?php the_title(); ?></h2><?php
             the_date('F j, Y','<p class="meta">Posted on ','</p>');
             the_content();
             the_tags('<p>Tags: ', ', ', '</p>'); ?>
-        </div><!-- /.span8 -->
-    </div><?php
-endwhile; // end of the loop.
+	    </div><?php
+	endwhile; // end of the loop.
+	get_sidebar('right');
 get_footer(); ?>