+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 Change Log - MOST
 ===
 
+### 0.1.8 - 13 Jan 2014
+* Fixed admin css form bugs
+* Updated `sidebar-left.php` and `sidebar-right.php` to be more flexible
+* Added sidebars to `page.php`, `single.php`, `single-event.php`, and `single-show.php`
+* Added one sidebar to `template-pages/page-calendar.php`
+* Built `search.php`, `404.php`, `archive.php`, and `index.php`
+* Added `inc/posts-widget.php` and built framework
+
+- - -
+
+### 0.1.7 - 10 Jan 2014
+* Bug fix: added white background to calendar rows
+
+- - -
+
 ### 0.1.6 - 09 Jan 2014
 * Restructured content for `single.php` and `page.php`
 * Added files `single-event.php` and `single-show.php` to display custom post content